# server/src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "공모아 API 문서",
    "description": "공모아 서버 API관련 문서입니다.",
    "version": "1.0.0"
  },
  "host": "https://gongmoa.site/api",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user_menu/userMenuList": {
      "get": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "사용자 메뉴 관리 - 리스트 불러오기 관련 로직",
        "description": "",
        "parameters": [
          {
            "name": "pg",
            "in": "query",
            "type": "string"
          },
          {
            "name": "pp",
            "in": "query",
            "type": "string"
          },
          {
            "name": "parentCode",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user_menu/setUserMenuStatus": {
      "post": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "사용자 메뉴 관리 - 상태변경 관련 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "example": "any"
                },
                "menu_code": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/user_menu/getUserMenuCode": {
      "get": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "사용자 메뉴 관리 - 부모 코드 불러오기",
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user_menu/getUserMenuDetail": {
      "get": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "사용자 메뉴 관리 - 상세 불러오기",
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user_menu/modifyUserMenu": {
      "post": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "사용자 메뉴 관리 - 수정관련 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "act": {
                  "example": "any"
                },
                "menu_code": {
                  "example": "any"
                },
                "menu_name": {
                  "example": "any"
                },
                "menu_path": {
                  "example": "any"
                },
                "use_yn": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/user_menu/setUserMenuOrder": {
      "post": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "사용자 메뉴 관리 - 순서변경 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "parentCode": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/userMenuList": {
      "get": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "사용자 메뉴 관리 - 리스트 불러오기 관련 로직",
        "description": "",
        "parameters": [
          {
            "name": "pg",
            "in": "query",
            "type": "string"
          },
          {
            "name": "pp",
            "in": "query",
            "type": "string"
          },
          {
            "name": "parentCode",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/setUserMenuStatus": {
      "post": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "사용자 메뉴 관리 - 상태변경 관련 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "example": "any"
                },
                "menu_code": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/getUserMenuCode": {
      "get": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "사용자 메뉴 관리 - 부모 코드 불러오기",
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/getUserMenuDetail": {
      "get": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "사용자 메뉴 관리 - 상세 불러오기",
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/modifyUserMenu": {
      "post": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "사용자 메뉴 관리 - 수정관련 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "act": {
                  "example": "any"
                },
                "menu_code": {
                  "example": "any"
                },
                "menu_name": {
                  "example": "any"
                },
                "menu_path": {
                  "example": "any"
                },
                "use_yn": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/setUserMenuOrder": {
      "post": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "사용자 메뉴 관리 - 순서변경 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "parentCode": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}